requirements.txt
check-manifest>=0.37
coverage>=4.5.2
Cython>=0.29.3
m2r
mistune==0.8.4; python_version <= '3.7'
# https://developercommunity.visualstudio.com/content/problem/1207405/fmod-after-an-update-to-windows-2004-is-causing-a.html
numpy
psutil
pyjls>=0.4.2
pyjoulescope_driver>=1.0.4
pymonocypher>=0.1.3
pywin32; sys_platform == 'win32'
# python {path_to_python}\scripts\pywin32_postinstall.py -install
python-dateutil>=2.7.3
recommonmark
sphinx>=3.1.0
wheel
